Fleshed out ValueSet some more, continued grader.

+++ src/redoku/generator.rs
@@ -1,9 +1,9 @@
 use extra::rand::Randomizer;
-use utils::{random_cell_value, read_u8_in_range};
 use grader::{Difficulty, RedokuGrader};
 use redoku::Redoku;
-use value::CellValue;
 use solver::RedokuSolver;
+use utils::{random_cell_value, read_u8_in_range};
+use value::Value;
 
 #[derive(Debug)]
 pub struct RedokuGenerator {
@@ -12,7 +12,7 @@ pub struct RedokuGenerator {
 
 impl RedokuGenerator {
     pub fn build(rand: &mut Randomizer, difficulty: Difficulty) -> Redoku {
-        let mut redoku = Redoku::new();
+        let mut redoku = Redoku::new(); // TODO: with_capacity based on difficulty
         let mut filled_cells = 0;
 
         loop {
@@ -23,14 +23,14 @@ impl RedokuGenerator {
                 let x = read_u8_in_range(rand, 0..9);
                 let y = read_u8_in_range(rand, 0..9);
 
-                if redoku[(x as usize, y as usize)].is_some() {
+                if redoku[(x, y)].is_some() {
                     continue;
                 }
 
                 loop {
                     let value = random_cell_value(rand);
 
-                    if redoku.place_if_valid(x as usize, y as usize, Some(value)) {
+                    if redoku.place_if_valid(x, y, Some(value)) {
                         filled_cells += 1;
                         break;
                     }
@@ -47,7 +47,6 @@ impl RedokuGenerator {
         }
 
         println!("{:?}", redoku);
-
 
         redoku
     }
